+++ L1C_Conversion_Pipeline/conversion_stub.py
@@ -1,11 +1,8 @@
-from caltrack_dict import list_caltrack, I_Q_U, LAT_LONG_ALT
+from L1C_Conversion_Pipeline.caltrack_dict import list_caltrack, LAT_LONG_ALT
 import h5py
 from pyhdf.SD import SD, SDC
 import os
 import numpy as np
-import copy
-from netCDF4 import Dataset
-
 
 HDF4 = '.hdf'
 HDF5 = '.h5'
